+++ app/Http/Controllers/TypeController.php
@@ -15,7 +15,11 @@ class TypeController extends Controller {
 
 
     public function index() {
-        return view('admin/admin_typetable');
+        if(session()->has('userID')) {
+            return view('admin/admin_typetable');
+        } else {
+            return view('admin/no_admin');
+        }
     }
 
     public function typetable() {
@@ -33,34 +37,45 @@ class TypeController extends Controller {
 
 
     public function edit($id) {
-        $data['title'] = "Editovať typ";
+        if(session()->has('userID')) {
+            $data['title'] = "Editovať typ";
 
-        $type = Type::select('*')->where("id_type","=",$id)->get()->first();
+            $type = Type::select('*')->where("id_type", "=", $id)->get()->first();
 
-        $data['type'] =  $type;
+            $data['type'] = $type;
 
-        return view('admin/type_edit',$data);
+            return view('admin/type_edit', $data);
+        } else {
+            return view('admin/no_admin');
+        }
     }
 
 
     public function delete($id) {
+        if(session()->has('userID')) {
+            if ($id == null) {
+                return abort(404);
+            }
+            $type = Type::find($id);
 
-        if($id == null) {
-            return abort(404);
+            $type->delete();
+            return redirect('/it-admin/types/');
+        } else {
+            return view('admin/no_admin');
         }
-        $type = Type::find($id);
-
-        $type->delete();
-        return redirect('/it-admin/types/');
     }
 
 
     public function edit_validator(Request $request) {
+        if($request->submit == "submit") {
+            $type = Type::where("id_type", "=", $request->input('id'))->first();
+            $type->update([
+                "title" => $request->input('title')]);
+            return redirect()->action('TypeController@index');
+        } else {
 
-        $type = Type::where("id_type","=",$request->input('id'))->first();
-        $type->update([
-            "title" => $request->input('title')]);
-        return redirect()-> action('TypeController@index');
+            return redirect()->action('TypeController@index');
+        }
     }
